+++ oggm/core/flowline.py
@@ -2190,13 +2190,13 @@ def flowline_model_run(gdir, output_filesuffix=None, mb_model=None,
 @entity_task(log)
 def run_random_climate(gdir, nyears=1000, y0=None, halfsize=15,
                        bias=None, seed=None, temperature_bias=None,
+                       precipitation_factor=None,
                        store_monthly_step=False,
                        climate_filename='climate_historical',
                        climate_input_filesuffix='',
                        output_filesuffix='', init_model_fls=None,
                        zero_initial_glacier=False,
-                       unique_samples=False,
-                       **kwargs):
+                       unique_samples=False, **kwargs):
     """Runs the random mass-balance model for a given number of years.
 
     This will initialize a
@@ -2224,6 +2224,10 @@ def run_random_climate(gdir, nyears=1000, y0=None, halfsize=15,
         be useful if you want to have the same climate years for all of them
     temperature_bias : float
         add a bias to the temperature timeseries
+    precipitation_factor: float
+        multiply a factor to the precipitation time series
+        default is None and means that the precipitation factor from the
+        calibration is applied which is cfg.PARAMS['prcp_scaling_factor']
     store_monthly_step : bool
         whether to store the diagnostic data at a monthly time step or not
         (default is yearly)
@@ -2258,6 +2262,8 @@ def run_random_climate(gdir, nyears=1000, y0=None, halfsize=15,
 
     if temperature_bias is not None:
         mb.temp_bias = temperature_bias
+    if precipitation_factor is not None:
+        mb.prcp_fac = precipitation_factor
 
     return flowline_model_run(gdir, output_filesuffix=output_filesuffix,
                               mb_model=mb, ys=0, ye=nyears,
@@ -2270,13 +2276,13 @@ def run_random_climate(gdir, nyears=1000, y0=None, halfsize=15,
 @entity_task(log)
 def run_constant_climate(gdir, nyears=1000, y0=None, halfsize=15,
                          bias=None, temperature_bias=None,
+                         precipitation_factor=None,
                          store_monthly_step=False,
                          output_filesuffix='',
                          climate_filename='climate_historical',
                          climate_input_filesuffix='',
                          init_model_fls=None,
-                         zero_initial_glacier=False,
-                         **kwargs):
+                         zero_initial_glacier=False, **kwargs):
     """Runs the constant mass-balance model for a given number of years.
 
     This will initialize a
@@ -2301,6 +2307,10 @@ def run_constant_climate(gdir, nyears=1000, y0=None, halfsize=15,
         to zero
     temperature_bias : float
         add a bias to the temperature timeseries
+    precipitation_factor: float
+        multiply a factor to the precipitation time series
+        default is None and means that the precipitation factor from the
+        calibration is applied which is cfg.PARAMS['prcp_scaling_factor']
     store_monthly_step : bool
         whether to store the diagnostic data at a monthly time step or not
         (default is yearly)
@@ -2328,6 +2338,8 @@ def run_constant_climate(gdir, nyears=1000, y0=None, halfsize=15,
 
     if temperature_bias is not None:
         mb.temp_bias = temperature_bias
+    if precipitation_factor is not None:
+        mb.prcp_fac = precipitation_factor
 
     return flowline_model_run(gdir, output_filesuffix=output_filesuffix,
                               mb_model=mb, ys=0, ye=nyears,
@@ -2344,7 +2356,8 @@ def run_from_climate_data(gdir, ys=None, ye=None, min_ys=None, max_ys=None,
                           climate_input_filesuffix='', output_filesuffix='',
                           init_model_filesuffix=None, init_model_yr=None,
                           init_model_fls=None, zero_initial_glacier=False,
-                          bias=None, **kwargs):
+                          bias=None, temperature_bias=None,
+                          precipitation_factor=None, **kwargs):
     """ Runs a glacier with climate input from e.g. CRU or a GCM.
 
     This will initialize a
@@ -2393,6 +2406,12 @@ def run_from_climate_data(gdir, ys=None, ye=None, min_ys=None, max_ys=None,
         bias of the mb model. Default is to use the calibrated one, which
         is often a better idea. For t* experiments it can be useful to set it
         to zero
+    temperature_bias : float
+        add a bias to the temperature timeseries
+    precipitation_factor: float
+        multiply a factor to the precipitation time series
+        default is None and means that the precipitation factor from the
+        calibration is applied which is cfg.PARAMS['prcp_scaling_factor']
     kwargs : dict
         kwargs to pass to the FluxBasedModel instance
     """
@@ -2428,6 +2447,11 @@ def run_from_climate_data(gdir, ys=None, ye=None, min_ys=None, max_ys=None,
     mb = MultipleFlowlineMassBalance(gdir, mb_model_class=PastMassBalance,
                                      filename=climate_filename, bias=bias,
                                      input_filesuffix=climate_input_filesuffix)
+
+    if temperature_bias is not None:
+        mb.temp_bias = temperature_bias
+    if precipitation_factor is not None:
+        mb.prcp_fac = precipitation_factor
 
     if ye is None:
         # Decide from climate (we can run the last year with data as well)